Use default AWS Credentials provider chain
so we do not have to use a hardcoded credentials/config file and we
reuse the default SDK logic to lookup credentials.

BREAKING CHANGE: Also change region lookup to a CLI parameter in order to remove the
dependency to the `.aws/config` which seems to be a feature of the AWS
CLI, but not the Java SDK.

Credentials stored in profiles should still work because the provider
chain includes the `ProfileCredentialsProvider`.